pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mintchat"
version = "2.60.6"
description = "多模态猫娘女仆智能体（LangChain/LangGraph + PyQt6）"
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [{ name = "MintChat Team" }]

dependencies = [
  # LangChain / Agent
  "langchain>=1.0.7",
  "langchain-core>=1.0.5",
  "langchain-community>=0.4.1",
  "langgraph>=1.0.0",
  "langchain-openai>=1.0.3",
  "langchain-anthropic>=1.0.2",
  "langchain-google-genai>=3.0.1",
  # OpenAI SDK
  "openai>=2.7.1",
  # Multimodal
  "pillow>=11.0.0",
  "opencv-python>=4.10.0",
  "soundfile>=0.12.1",
  "pydub>=0.25.1",
  "imageio-ffmpeg>=0.4.9",
  "librosa>=0.10.2",
  "pytesseract>=0.3.13",
  "pygame>=2.6.1",
  # Vector DB / Retrieval
  "chromadb>=1.3.4,<2.0.0",
  "langchain-chroma>=1.0.0",
  "faiss-cpu>=1.9.0",
  "sentence-transformers>=5.1.2",
  # PyTorch (CUDA 13.0 on Windows/Linux; CPU on macOS)
  "torch==2.9.1+cu130; sys_platform != 'darwin'",
  "torch==2.9.1; sys_platform == 'darwin'",
  # Cache
  "redis>=5.0.0",
  # Config / HTTP / Logging
  "pydantic>=2.12.4",
  "pydantic-settings>=2.11.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "aiohttp>=3.13.2",
  "httpx>=0.27.0",
  "loguru>=0.7.3",
  "psutil>=6.1.0",
  # MCP
  "mcp>=1.21.1",
  "mcp-server-time>=2025.9.25",
  "starlette>=0.50.0",
  "sse-starlette>=3.0.3",
  "pywin32>=311; sys_platform == 'win32'",
  # GUI
  "PyQt6>=6.8.0",
  "live2d-py>=0.5.4",
  # Observability
  "langsmith>=0.2.11",
  "funasr>=1.2.9",
]

[project.optional-dependencies]
# ASR (语音识别) 依赖较重，且部分依赖在 Windows + Python 3.13 下可能需要本地编译工具。
# 如需启用 ASR，请执行：`uv sync --locked --no-install-project --extra asr`
asr = [
  "funasr>=1.2.9",
  # torchaudio is imported by FunASR at runtime (audio loading backend)
  "torchaudio==2.9.1+cu130; sys_platform != 'darwin'",
  "torchaudio==2.9.1; sys_platform == 'darwin'",
  "modelscope>=1.19.0",
  "huggingface-hub>=0.25.0",
  "transformers>=4.46.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=5.0.0",
  "black>=24.10.0",
  "flake8>=7.1.1",
  "mypy>=1.13.0",
  "ipython>=8.30.0",
  "jupyter>=1.1.1",
]

[tool.uv]
# 不默认安装 dev 组，避免运行环境拉入 black/flake8 等开发依赖。
default-groups = []

[tool.uv.sources]
# Use PyTorch CUDA 13.0 wheels on Windows/Linux; fall back to PyPI (CPU) on macOS.
torch = { index = "pytorch-cu130", marker = "sys_platform != 'darwin'" }
torchaudio = { index = "pytorch-cu130", marker = "sys_platform != 'darwin'" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.setuptools.packages.find]
# 代码当前以 `src/` 作为顶层包（import 路径为 `src.*`）
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
no_implicit_optional = false

# Mypy is enforced in CI, but this codebase is still incrementally adopting typing.
# For now, keep mypy focused on configuration/version modules. The remaining subsystems are highly dynamic
# (PyQt, LangChain tools, multimodal backends, etc.) and will be typed incrementally.
[[tool.mypy.overrides]]
module = ["src.gui", "src.gui.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["src.agent", "src.agent.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["src.utils", "src.utils.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["src.auth", "src.auth.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["src.multimodal", "src.multimodal.*"]
ignore_errors = true

# Incremental typing: enable mypy for key agent modules that are now guarded by tests
# and have stable, low-dynamic surfaces.
[[tool.mypy.overrides]]
module = ["src.agent.advanced_memory", "src.agent.memory_retriever"]
ignore_errors = false
